+++ common/src/main/java/cc/cassian/cauldrons/compat/emi/DippingEmiRecipe.java
@@ -15,16 +15,16 @@ import net.minecraft.world.item.crafting.RecipeHolder;
 
 public class DippingEmiRecipe extends BasicEmiRecipe {
 
-    private final ItemStack potionForDisplay;
+    private final EmiStack potionForDisplay;
 
     public DippingEmiRecipe(RecipeHolder<DippingRecipe> recipe, RegistryAccess registryAccess) {
         super(CauldronModEmiPlugin.DIPPING_CATEGORY, recipe.id(), 100, 18);
         // reagent
         inputs.add(EmiIngredient.of(recipe.value().getReagent()));
         // potion item
-        var potion = PotionContents.createItemStack(Items.POTION, recipe.value().getPotion());
-        inputs.add(EmiStack.of(potion));
-        this.potionForDisplay = PotionContents.createItemStack(CauldronModItems.CAULDRON_CONTENTS.get(), recipe.value().getPotion());
+        var input = CauldronModEmiPlugin.getResultForDisplay(recipe.value().getPotion());
+        inputs.add(EmiIngredient.of(input.getA().getEmiStacks()));
+        potionForDisplay = input.getB();
         // output
         var result = recipe.value().getResultItem(registryAccess);
         outputs.add(EmiStack.of(result));
@@ -34,7 +34,7 @@ public class DippingEmiRecipe extends BasicEmiRecipe {
     public void addWidgets(WidgetHolder widgetHolder) {
         widgetHolder.addSlot(inputs.getFirst(), 0, 0);
         widgetHolder.addTexture(EmiTexture.PLUS, 20, 2);
-        widgetHolder.addSlot(EmiStack.of(potionForDisplay), 36, 0);
+        widgetHolder.addSlot(potionForDisplay, 36, 0);
         widgetHolder.addTexture(EmiTexture.FULL_ARROW, 56, 1);
         widgetHolder.addSlot(getOutputs().getFirst(), 81, 0).recipeContext(this);