+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Particles and heat requirements are now part of the recipe format.
+- Brewing input and results no longer need to be a potion, specifying `id: "minecraft:lava_cauldron` for your `potion` or `result` will use that block as the result.
 
 ### Fixed
 - Duplicated item list in WTHIT/Jade.